requirements.txt
requests~=2.31.0
pygame~=2.5.2

pillow~=10.3.0